refactor: consolidate streams into responses/streams.py
The code is actually quite simple and fits comfortably in one file,
there isn't a need for a whole streams module. It makes more sense imo
for streams to be a sub concept in responses.
