+++ python/mirascope/llm/__init__.py
@@ -19,7 +19,6 @@ from . import (
     models,
     prompts,
     responses,
-    streams,
     tools,
     types,
 )
@@ -72,16 +71,23 @@ from .responses import (
     AsyncContextResponse,
     AsyncContextStreamResponse,
     AsyncResponse,
+    AsyncStream,
     AsyncStreamResponse,
+    AsyncTextStream,
+    AsyncThoughtStream,
+    AsyncToolCallStream,
     ChunkIterator,
     ContextResponse,
     ContextStreamResponse,
     FinishReason,
     Response,
+    Stream,
     StreamResponse,
     StreamResponseChunk,
+    TextStream,
+    ThoughtStream,
+    ToolCallStream,
 )
-from .streams import AsyncStream, Stream
 from .tools import (
     AsyncContextTool,
     AsyncContextToolkit,
@@ -107,7 +113,10 @@ __all__ = [
     "AsyncResponse",
     "AsyncStream",
     "AsyncStreamResponse",
+    "AsyncTextStream",
+    "AsyncThoughtStream",
     "AsyncTool",
+    "AsyncToolCallStream",
     "AsyncToolkit",
     "Audio",
     "AuthenticationError",
@@ -145,16 +154,19 @@ __all__ = [
     "TextChunk",
     "TextEndChunk",
     "TextStartChunk",
+    "TextStream",
     "Thought",
     "ThoughtChunk",
     "ThoughtEndChunk",
     "ThoughtStartChunk",
+    "ThoughtStream",
     "TimeoutError",
     "Tool",
     "ToolCall",
     "ToolCallChunk",
     "ToolCallEndChunk",
     "ToolCallStartChunk",
+    "ToolCallStream",
     "ToolNotFoundError",
     "ToolOutput",
     "Toolkit",
@@ -176,7 +188,6 @@ __all__ = [
     "prompt",
     "prompts",
     "responses",
-    "streams",
     "tool",
     "tools",
     "types",
