+++ python/mirascope/llm/responses/__init__.py
@@ -16,13 +16,27 @@ from .stream_response import (
     ContextStreamResponse,
     StreamResponse,
 )
+from .streams import (
+    AsyncStream,
+    AsyncTextStream,
+    AsyncThoughtStream,
+    AsyncToolCallStream,
+    Stream,
+    TextStream,
+    ThoughtStream,
+    ToolCallStream,
+)
 
 __all__ = [
     "AsyncChunkIterator",
     "AsyncContextResponse",
     "AsyncContextStreamResponse",
     "AsyncResponse",
+    "AsyncStream",
     "AsyncStreamResponse",
+    "AsyncTextStream",
+    "AsyncThoughtStream",
+    "AsyncToolCallStream",
     "ChunkIterator",
     "ContextResponse",
     "ContextStreamResponse",
@@ -31,7 +45,11 @@ __all__ = [
     "RawChunk",
     "Response",
     "RootResponse",
+    "Stream",
     "StreamResponse",
     "StreamResponseChunk",
+    "TextStream",
+    "ThoughtStream",
+    "ToolCallStream",
     "_utils",
 ]
